separated types.rs into many other submodules inside types/
